Additional changes for OpenBSD support

+++ health/health.c
@@ -481,8 +481,8 @@ void *health_main(void *ptr) {
 		if (unlikely(check_if_resumed_from_suspention())) {
 			apply_hibernation_delay = 1;
 
-			info("Postponing alarm checks for %ld seconds, because it seems that the system was just resumed from suspension.",
-				 hibernation_delay
+			info("Postponing alarm checks for %lld seconds, because it seems that the system was just resumed from suspension.",
+				 (long long)hibernation_delay
 			);
 		}
 
@@ -504,7 +504,9 @@ void *health_main(void *ptr) {
 
 			if (unlikely(apply_hibernation_delay)) {
 
-				info("Postponing health checks for %ld seconds, on host '%s'.", hibernation_delay, host->hostname
+				info("Postponing health checks for %lld seconds, on host '%s'.",
+                        (long long)hibernation_delay,
+                        host->hostname
 				);
 
 				host->health_delay_up_to = now + hibernation_delay;
